Fixed and finalized projectile impacts ( for now )
Somehow disabling CCD doesn't affect our projectile hit accuracy, and actually improves accuracy agaisnt world actors. Added the functionality to determine how much impulse is applied on world actors. This is a miracle of engineering and I honestly don't know how it is working this way because it shouldn't, but it is. Henrique the coder tyvm ladies n gents