More work on child actor components
Finalized (hopefully) the base turret blueprint, which will be used as a customizable child actor component inside the vehicle blueprint. Still need to add some variables such as rate of fire and overheat times, etc. Still need to do this same proccess for melee weapons.
Made the turret rotate smoothly as well, instead of snapping to the direction of our camera.