Child Actors Part 5
Every issue resolved, created data struct that holds the vehicle customization information (turret, melee, colour [in the future]) and placed 2 vehicles in my level with different values on the data struct as proof of concept. Changing these components at runtime has been proven to be possible as well, yet not implemented. However, you can press GHJ to possess the 3 different vehicles in my level, and it's sort of built on the same concept.