Merged everyone's work
Changed the vehicle to have Kamil's movement system, created new SaveGame blueprint named VehicleSettingsSave, which is saved under the slot of the same name, which hold the VehicleSettings data struct. Nik's customization changed to allow for turret and melee choice, leading into the Quarry v2 level, with a new GameMode blueprint that uses the saved data to spawn our final vehicle with the correct customized parts.