New Spring Arm Component Perfected
Managed to fix what I wanted, when coming out of a collision the camera now pans out until it reaches its max distance, instead of snapping to max distance.

Speed of transition adjustable with the float var "LengthAdjustSpeed", which in the future will also be in use when changing the max distance of the camera in relation to the vehicle's speed.