Reworking the vehicle initialisation
This was apparently needed to pass more information such as the turret rotation and shooting. Still not finished, but the logic is in a better place and runs more smoothly. At the moment, players are still unable to see other players' turret rotating and connected clients will experience bugged turret rotation still.