Removed car fire vfx from spawning upon death
I made a choice to remove it as it was unoptimised, causing a lot of framerate issues. The fire was also not properly set up with moving objects in mind, so when the car moved the vfx would just teleport along with it, which appeared very broken and would just take away from the user experience instead of enhancing it.